refactor: move up http inner classes from gql helper
